Класс Startup налажен.
