+++ src/TelegramBot/Configurations/BotConfiguration.cs
@@ -1,0 +1,7 @@
+﻿namespace ThursdayMeetingBot.TelegramBot.Configurations
+{
+    /// <summary>
+    ///     Configurations for bot
+    /// </summary>
+    internal record BotConfiguration(string AccessToken);
+}
