+++ src/TelegramBot/Constants/HttpClientConstants.cs
@@ -1,0 +1,8 @@
+﻿namespace ThursdayMeetingBot.TelegramBot.Constants
+{
+    internal static class HttpClientConstants
+    {
+        internal const string Name = "TelegramApi";
+        internal const string UriString = "https://api.telegram.org";
+    }
+}
