--- src/ThursdayMeetingBot/ThursdayMeetingBot.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3"/>
-    </ItemGroup>
-
-</Project>
